PythonDocker/requirements.txt
flask
gunicorn
tensorflow-cpu==2.5.0
numpy
pandas